requirements.txt
mss
Pillow
transformers
torch
accelerate